EduTrack-main/backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
pymongo==4.10.1
mysql-connector-python==9.2.0
bcrypt==4.2.0
python-dotenv==1.0.1
PyJWT==2.10.1
